requirements.txt
pytest
pylint
black
flask
gunicorn
numpy
google.cloud
google.auth
google-cloud-automl
json
